fixed code climate fixes
