--- components/src/components/SimpleAddressSelect/editForm/Component.edit.validation.ts
@@ -1,4 +1,0 @@
-import common from '../../Common/Simple.edit.validation';
-export default [
-    ...common,
-];
