--- components/src/components/SimpleMultiUseAddressSelect/editForm/Component.edit.display.ts
@@ -1,9 +1,0 @@
-import common from '../../Common/Simple.edit.display';
-import commonv2 from '../../Common/Simple.edit.display.v2'
-export default {
-  key: 'display',
-  components: [
-    ...common,
-    ...commonv2
-  ]
-}
