--- components/src/components/SimplePostalCode/editForm/Component.edit.data.ts
@@ -1,7 +1,0 @@
-import common from '../../Common/Simple.edit.data';
-import radioData from '../../Common/simple.radio.data';
-
-export default [
-  ...common,
-  radioData
-]
