+++ components/src/components/address/SimpleAddressSelect/Component.form.ts
@@ -1,6 +1,6 @@
 import baseEditForm from 'formiojs/components/_classes/component/Component.form';
 import EditDisplay from './editForm/Component.edit.display';
-import FormExtension from '../Common/Component.edit.data.extension';
+import FormExtension from '../../Common/Component.edit.data.extension';
 
 export default function(...extend) {
     return baseEditForm([
