+++ components/src/components/address/SimpleAddressSelect/editForm/Component.edit.display.ts
@@ -1,4 +1,4 @@
-import common from '../../Common/Simple.edit.display';
+import common from '../../../Common/Simple.edit.display';
 export default {
   key: 'display',
   components: [
