+++ components/src/components/address/SimpleAddressSelect/editForm/Component.edit.validation.ts
@@ -1,0 +1,4 @@
+import common from '../../../Common/Simple.edit.validation';
+export default [
+    ...common,
+];
