+++ components/src/components/address/SimpleCountry/Component.ts
@@ -1,31 +1,22 @@
 import {Components} from 'formiojs';
 const ParentCompoenent = (Components as any).components.select
-import { Constants } from '../Common/Constants';
+import { Constants } from '../../Common/Constants';
 import editForm from './Component.form';
 
-const ID = 'multiuseaddrselect';
-const DISPLAY = 'MultiUseAddrselect';
+const ID = 'country';
+const DISPLAY = 'Country';
 
 export default class Component extends (ParentCompoenent as any){
 
   static schema (...extend){
         return ParentCompoenent.schema({
           label: DISPLAY,
-          type: ID,
           key: ID,
-          clearOnHide: true,
-          tableView: true,
-          persistent: false,
-          autoAdjust: true,
-          hideLabel: true, 
-        template: '<span>{{ item.value }}</span>',
-          clearOnRefresh: false,
-          limit: 100,
-         
-        validate: {
-          required: true
-        },
-        dataSrc: 'url',       
+          disabled:true,
+          defaultValue:'Canada',
+          type: 'select',
+          input: true,
+             
         }, ...extend)
     }
 
@@ -33,7 +24,7 @@ export default class Component extends (ParentCompoenent as any){
     static get builderInfo() {
         return {
           title: DISPLAY,
-          icon: 'building',
+          icon: 'flag',
           group: 'simple',
           documentation: Constants.DEFAULT_HELP_LINK,
           weight: 30,
