+++ components/src/components/address/SimpleMultiUseAddressSelect/Component.ts
@@ -1,22 +1,31 @@
 import {Components} from 'formiojs';
 const ParentCompoenent = (Components as any).components.select
-import { Constants } from '../Common/Constants';
+import { Constants } from '../../Common/Constants';
 import editForm from './Component.form';
 
-const ID = 'country';
-const DISPLAY = 'Country';
+const ID = 'multiuseaddrselect';
+const DISPLAY = 'MultiUseAddrselect';
 
 export default class Component extends (ParentCompoenent as any){
 
   static schema (...extend){
         return ParentCompoenent.schema({
           label: DISPLAY,
+          type: ID,
           key: ID,
-          disabled:true,
-          defaultValue:'Canada',
-          type: 'select',
-          input: true,
-             
+          clearOnHide: true,
+          tableView: true,
+          persistent: false,
+          autoAdjust: true,
+          hideLabel: true, 
+        template: '<span>{{ item.value }}</span>',
+          clearOnRefresh: false,
+          limit: 100,
+         
+        validate: {
+          required: true
+        },
+        dataSrc: 'url',       
         }, ...extend)
     }
 
@@ -24,7 +33,7 @@ export default class Component extends (ParentCompoenent as any){
     static get builderInfo() {
         return {
           title: DISPLAY,
-          icon: 'flag',
+          icon: 'building',
           group: 'simple',
           documentation: Constants.DEFAULT_HELP_LINK,
           weight: 30,
