+++ components/src/components/address/SimplePostalCode/Component.form.ts
@@ -1,7 +1,7 @@
 import baseEditForm from 'formiojs/components/_classes/component/Component.form';
 import EditDisplay from './editForm/Component.edit.display';
 import EditValidation from './editForm/Component.edit.validation';
-import FormExtension from '../Common/Component.edit.data.extension';
+import FormExtension from '../../Common/Component.edit.data.extension';
 
 export default function(...extend) {
     return baseEditForm([
