+++ components/src/components/address/SimplePostalCode/editForm/Component.edit.validation.ts
@@ -1,4 +1,4 @@
-import common from '../../Common/Simple.edit.validation';
+import common from '../../../Common/Simple.edit.validation';
 export default [
     ...common,
     {
